+++ src/pages/login.tsx
@@ -5,6 +5,8 @@ import IdAndPasswordAuth from "@/components/auth/IdAndPasswordAuth";
 import { useRouter } from "next/router";
 import GitHubButton from "@/components/ui/button/social/GitHubButton";
 import GitHubAuth from "@/components/auth/GitHubAuth";
+import { GithubAuthProvider, signInWithPopup } from "firebase/auth";
+import { auth } from "../../firebase";
 
 const Wrapper = styled.div`
   /* height: 100%; */
@@ -124,6 +126,7 @@ const SocialBox = styled.div`
     display: block;
   } */
 `;
+const provider = new GithubAuthProvider();
 
 export default function Login() {
   const emailIdRef = useRef<HTMLDivElement>(null);
@@ -154,6 +157,16 @@ export default function Login() {
       LoginFormRef.current.classList.add("hidden");
     }
   };
+  const handleGitHubLogin = async () => {
+    try {
+      const userCred = await signInWithPopup(auth, provider);
+      const user = userCred.user;
+      console.log("GitHub Login: ", user);
+      router.push("/");
+    } catch (error) {
+      console.log("🚨 GitHub Login Error: ", error);
+    }
+  };
 
   return (
     <Wrapper>
@@ -165,7 +178,7 @@ export default function Login() {
           이메일 링크
         </LoginBox>
         <LoginBox>
-          <SocialBox>
+          <SocialBox onClick={handleGitHubLogin}>
             <GitHubButton />
           </SocialBox>
         </LoginBox>
